requirements.txt
# # Core Framework
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# pydantic==2.5.0
# pydantic-settings==2.1.0

# # Kafka
# kafka-python==2.0.2
# confluent-kafka==2.3.0

# # WebSocket
# websockets==12.0
# python-socketio==5.11.0

# # Authentication & Security
# python-jose[cryptography]==3.3.0
# python-multipart==0.0.6
# passlib[bcrypt]==1.7.4

# # Database & Caching
# redis==5.0.1
# sqlalchemy==2.0.23
# alembic==1.13.0

# # Monitoring & Observability
# prometheus-client==0.19.0
# opentelemetry-api==1.21.0
# opentelemetry-sdk==1.21.0
# opentelemetry-instrumentation-fastapi==0.42b0
# # opentelemetry-instrumentation-kafka==0.42b0
# opentelemetry-exporter-prometheus==1.12.0rc1

# # Circuit Breaker
# # circuit-breaker==1.0.0
# tenacity==8.2.3

# # Utilities
# python-dotenv==1.0.0
# structlog==23.2.0
# colorlog==6.8.0
# pytz==2023.3
# croniter==2.0.1

# # Development
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2
# pytest-mock==3.12.0

# # JSON Schema
# jsonschema==4.20.0

# # Async support
# aiofiles==23.2.1
# asyncio-mqtt==0.16.1
# aiohttp
# asyncpg
# aioredis
# authlib

# # Core Framework
# fastapi==0.115.1
# uvicorn[standard]==0.24.0
# pydantic==2.7.4
# pydantic-settings==2.8.1

# # Kafka
# kafka-python==2.0.2
# confluent-kafka==2.3.0

# # WebSocket
# websockets==12.0
# python-socketio==5.11.0

# # Authentication & Security
# python-jose[cryptography]==3.3.0
# python-multipart==0.0.6
# passlib[bcrypt]==1.7.4

# # Database & Caching
# redis==5.0.1
# sqlalchemy==2.0.23
# alembic==1.13.0

# # Monitoring & Observability
# prometheus-client==0.19.0
# opentelemetry-api
# opentelemetry-sdk==1.33.1
# opentelemetry-instrumentation-fastapi

# opentelemetry-exporter-prometheus==0.56b0


# # Circuit Breaker
# # circuit-breaker==1.0.0
# tenacity==8.2.3

# # Utilities
# python-dotenv==1.0.0
# structlog==23.2.0
# colorlog==6.8.0
# pytz==2023.3
# croniter==2.0.1

# # Development
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.27.0
# pytest-mock==3.12.0

# # JSON Schema
# jsonschema==4.20.0

# # Async support
# aiofiles==23.2.1
# asyncio-mqtt==0.16.1
# aiohttp
# asyncpg
# aioredis
# authlib

# # LangChain-related
# langchain==0.3.24
# langchain-core==0.3.56
# langchain-community==0.3.23
# chromadb==1.0.10

# Core Framework
fastapi==0.115.9
uvicorn[standard]==0.24.0
pydantic==2.7.4
pydantic-settings==2.8.1

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# WebSocket
websockets==12.0
python-socketio==5.11.0

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Database & Caching
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.0

# Monitoring & Observability
#prometheus-client==0.19.0
#opentelemetry-api==1.33.1
prometheus-client
opentelemetry-api
#opentelemetry-sdk>=1.33.0,<1.34
#opentelemetry-instrumentation-fastapi>=0.54b0,<0.55
opentelemetry-sdk
opentelemetry-instrumentation-fastapi

opentelemetry-exporter-prometheus==0.56b0


# Circuit Breaker
# circuit-breaker==1.0.0
tenacity==8.2.3

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
colorlog==6.8.0
pytz==2023.3
croniter==2.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.27.0
pytest-mock==3.12.0

# JSON Schema
jsonschema==4.20.0

# Async support
aiofiles==23.2.1
asyncio-mqtt==0.16.1
aiohttp
asyncpg
aioredis
authlib

# LangChain-related
langchain==0.3.24
langchain-core==0.3.56
langchain-community==0.3.23
chromadb==1.0.10